pull out lbry volume class
